deleted .rst files/angle buttons now incr & decr
