--- singlecrystaldiffractGUI/AUTHORS.rst
@@ -1,13 +1,0 @@
-=======
-Credits
-=======
-
-Maintainer
-----------
-
-* BNL-SULI-Sabrina Kay Sullens  <sabrina.kayg96@gmail.com>
-
-Contributors
-------------
-
-None yet. Why not be the first? See: CONTRIBUTING.rst
