+++ singlecrystaldiffractGUI/crystalxrayinterface/EmptyWindow.py
@@ -1,0 +1,17 @@
+import pygame
+pygame.init()
+
+width, height = 800, 600
+window = pygame.display.set_mode((width, height))
+pygame.display.set_caption("Hellow Window")  # Set the window title
+
+running = True
+while running:
+    for event in pygame.event.get():
+        if event.type == pygame.QUIT:
+            running = False
+
+    window.fill((255, 255, 255))
+
+    pygame.display.update()
+pygame.quit()
